Moved detect_video() back to yolo.py for now.  Variable name and old code cleanup.
