+++ yolo_video.py
@@ -1,59 +1,11 @@
 import sys
 import cv2
 import argparse
-from yolo import YOLO, yolo_parser
-
-def detect_video(yolo, video_path, output_path=""):
-    if output_path != "":
-        isOutput = True
-        print("Output Video Path: " , type(output_path))
-    else:
-        isOutput = False
-    vid = cv2.VideoCapture(video_path)
-    if not vid.isOpened():
-        raise IOError("Couldn't open webcam or video")
-    video_FourCC    = int(vid.get(cv2.CAP_PROP_FOURCC))
-    video_fps       = vid.get(cv2.CAP_PROP_FPS)
-    video_size      = (int(vid.get(cv2.CAP_PROP_FRAME_WIDTH)),
-                        int(vid.get(cv2.CAP_PROP_FRAME_HEIGHT)))
-    if isOutput:
-        print("!!! TYPE:", type(output_path), type(video_FourCC), type(video_fps), type(video_size))
-        out = cv2.VideoWriter(output_path, video_FourCC, video_fps, video_size)
-    accum_time = 0
-    curr_fps = 0
-    fps = "FPS: ??"
-    prev_time = timer()
-    while True:
-        return_value, frame = vid.read()
-        image = Image.fromarray(frame)
-        image = yolo.detect_image(image)
-        result = np.asarray(image)
-        curr_time = timer()
-        exec_time = curr_time - prev_time
-        prev_time = curr_time
-        accum_time = accum_time + exec_time
-        curr_fps = curr_fps + 1
-        if accum_time > 1:
-            accum_time = accum_time - 1
-            fps = "FPS: " + str(curr_fps)
-            curr_fps = 0
-        cv2.putText(result, text=fps, org=(3, 15), fontFace=cv2.FONT_HERSHEY_SIMPLEX,
-                    fontScale=0.50, color=(255, 0, 0), thickness=2)
-        cv2.namedWindow("result", cv2.WINDOW_NORMAL)
-        cv2.imshow("result", result)
-        if isOutput:
-            out.write(result)
-        if cv2.waitKey(1) & 0xFF == ord('q'):
-            break
-    yolo.close_session()
+from yolo import YOLO, yolo_parser, detect_video
 
 
-PARSED_ARGS = None
-'''
-if len(sys.argv) < 2:
-    print("Usage: $ python {0} [video_path] [output_path(optional)]", sys.argv[0])
-    exit()
-'''
+FLAGS = None
+
 if __name__ == '__main__':
     parser = yolo_parser()
 
@@ -66,6 +18,6 @@ if __name__ == '__main__':
       help='[optional] output video path'
     )
 
-    PARSED_ARGS = parser.parse_args()
+    FLAGS = parser.parse_args()
 
-    detect_video(YOLO(**vars(PARSED_ARGS)), PARSED_ARGS.video_path, PARSED_ARGS.output)
+    detect_video(YOLO(**vars(FLAGS)), FLAGS.video_path, FLAGS.output)
